change int to double (var)

+++ Main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 void main (){
-	int var = 0;
+	double var = 0;
 	cout << "helloWorld";
 	cin >> var;
 	cout << "\nYour var = " << var;
